requirements.txt
jsonschema>=4.20
requests>=2.31
PyYAML>=6.0
